Ошибки исправил

+++ lab/lab1/b/ArrayMass.java
@@ -11,42 +11,42 @@ import java.util.Scanner;
 public class ArrayMass {
     static Scanner input = new Scanner(System.in);
     public static void main (String [] args){
-        byte [] a = inputArray();
-        byte [] b = inputArray();
+        long [] a = inputArray();
+        long [] b = inputArray();
         printArray(mutippleArrays(a,b));
     }
-    static byte [] mutippleArrays(byte[] firstArray,byte[] secondArray){
+    static long [] mutippleArrays(long[] firstArray,long[] secondArray){
         int resultNumber = convertToNumberArray(firstArray) * convertToNumberArray(secondArray);
         return convertNumberToArray(resultNumber);
     }
-    static int convertToNumberArray(byte[] array){
+    static int convertToNumberArray(long[] array){
         int number = 0;
-        for (int i= array.length-1; i >= 0; i--){
+        for (int i = array.length - 1; i >= 0; i--) {
             number += array[i] * (int) Math.pow(10,array.length - i-1);
         }
         return number;
     }
-    static byte [] convertNumberToArray(int number){
+    static long [] convertNumberToArray(int number){
         int arrayLength = ("" + number).length();
-        byte [] result = new byte [arrayLength];
+        long [] result = new long [arrayLength];
         for (int i= arrayLength-1; i >= 0; i--){
-            result[i] = (byte) (number%10);
+            result[i] = (long) (number%10);
             number/=10;
         }
         return result;
     }
 
-    static byte[] inputArray(){
+    static long[] inputArray(){
         System.out.print("Введите длину массива:");
         int arrayLength = input.nextInt();
-        byte [] array = new byte [arrayLength];
+        long [] array = new long [arrayLength];
         System.out.print("Введите элементы массива:");
         for(int i=0; i< arrayLength; i++){
             array[i] = input.nextByte();
         }
         return array;
     }
-    static void printArray(byte [] array){
+    static void printArray(long [] array){
         System.out.print("[");
         for (int i=0; i< array.length; i++){
             System.out.print(array[i]+",");
